reddit_bot/requirements.txt
beautifulsoup4==4.11.1
black==22.3.0
certifi==2021.10.8
mypy-extensions==0.4.3
pathspec==0.9.0
praw==7.6.0
prawcore==2.3.0
python-dotenv==0.20.0
requests==2.27.1
schedule==1.1.0
typing_extensions==4.2.0
urllib3==1.26.9